backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=1.4
pymysql
python-dotenv
redis
passlib[bcrypt]
pydantic
typing-extensions
python-jose[cryptography]
python-multipart
